fixed a typing cast that caused an error when making a linux daemon

+++ src/rpcmining.cpp
@@ -542,8 +542,8 @@ Value getblocktemplate(const Array& params, bool fHelp)
         entry.push_back(Pair("depends", deps));
 
         int index_in_template = i - 1;
-        entry.push_back(Pair("fee", pblocktemplate->vTxFees[index_in_template]));
-        entry.push_back(Pair("sigops", pblocktemplate->vTxSigOps[index_in_template]));
+        entry.push_back(Pair("fee", (int64_t)pblocktemplate->vTxFees[index_in_template]));
+        entry.push_back(Pair("sigops", (int64_t)pblocktemplate->vTxSigOps[index_in_template]));
 
         transactions.push_back(entry);
     }
